Bugfix and Change
FIX: Fixed a bug in the service class, where a recipe id was used to retrieve a user by id. changed this so that the user is retrieved by username of the current logged in user.
CHANGE: as a result of the fix above, a recipe is now always created for the current logged in user.