COOK-153 - Add Expcetion Page
ADD: Controller that handles all the exceptions that occur in the application.
ADD: tiles definition for the exception. In addition to render a specific message for the user, a html comment is rendered that contains the stack trace (for dev)
ADD: added a exception message that is displayed for any nonspecific exception.