COOK-161 - general work
REFACTOR: if a non-authorized user tries to access the /account/detail page a UserNotAuthorized exception is thrown, since there is no page for an non-authorized user. This behaviour is used, because the url is protected and only accessible for authenticated users in general, which results in a redirect to the login page anyways.